refactor(api): handlers_recording.go 按域拆分 — 1443 行 → 4 个按域文件 (#611)
纯移动无逻辑变化：handlers_recording.go(602, 列表/时间线/AI 回写+路由注册) + handlers_recording_delete.go(131, 单删+批量删) + handlers_recording_frames.go(502, 下载/帧列表/AVI 帧/timelapse 帧/合并回放) + handlers_recording_media_meta.go(351, codec 探测缓存+帧时间戳/EXIF 解析)
